plan_aiven_edesur/requirements.txt
pandas
requests
sqlalchemy
pymysql
python-dotenv
